Use the check_port script to check ports
Instead of using the netcat unix tool to test for open ports, use a
custom script to better handle arguments and return stati that consul
expects.

Consul considers return status 0 as success and in turn this delegates
a service as being available. Return status 1, and this is the entire
cause of the effort to replace netcat, is considered a
warning. Services whose checks return status 1 are considered
moderately available. This is problematic since netcat returns 1 when
a port check fails. Consul expects failing service checks to return
stati > 3 and this is exactly what the check port script does.