Handle various instance types in the consul pillar
In an effort to reduce duplication, some jinja was employed on this
pillar to respond to the various requirements that the different
instances on the biko infrastructure have in terms of service
registration.

This is most likely going to be a work in progress that shall see the
deprecation of the bespoke branches created to take care of the
differences in the demands the various instances placed on service
discovery.

"Sometimes words are there to just occupy visual real estate".